+++ shared/src/types/transcript.types.js
@@ -1,3 +1,0 @@
-"use strict";
-// Defines shared types related to audio transcription results
-Object.defineProperty(exports, "__esModule", { value: true });